backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
pandas>=1.3.0
python-dotenv>=0.19.0
python-multipart>=0.0.5
